tinyint removed

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -32,7 +32,7 @@ class CreateUsersTable extends Migration
             $table->timestamps();
             $table->index('encadrant_id');
             $table->string('avatar',100)->default('default.png');
-            $table->tinyint('validation')->default(0);
+            $table->integer('validation')->default(0);
         });
     }
